+++ lib/src/models/cron.g.dart
@@ -7,6 +7,7 @@ part of 'cron.dart';
 // **************************************************************************
 
 Cron _$CronFromJson(Map<String, dynamic> json) => Cron(
+      name: json['name'] as String?,
       dayOfWeek: json['day_of_week'] as String,
       dayOfMonth: json['day_of_month'] as String,
       hour: json['hour'] as String,
@@ -14,6 +15,7 @@ Cron _$CronFromJson(Map<String, dynamic> json) => Cron(
     );
 
 Map<String, dynamic> _$CronToJson(Cron instance) => <String, dynamic>{
+      'name': instance.name,
       'day_of_week': instance.dayOfWeek,
       'day_of_month': instance.dayOfMonth,
       'hour': instance.hour,